added new pages and updated email sending

+++ src/app/routes/admin.py
@@ -39,11 +39,21 @@ def admin_users_page(page=1):
         amount_of_users = users.count()
     return render_template('Admin/users.html', users=users, page=page, total_pages=total_pages, search_query=search_query, amount_of_users=amount_of_users, form=form)
     
+@app.route('/admin/user/<int:id>', methods=['GET','POST'])
+@admin_required
+def admin_user_info_page(id):
+    user = User.query.get(id)
+    return render_template('Admin/user-info.html', user=user)
+
 @app.route('/admin/user/edit/<int:id>', methods=['GET','POST'])
 @admin_required
 def admin_user_edit_page(id):
     user = User.query.get(id)
     if request.method == 'POST':
+
+        if user.role == 'owner' and current_user.role != 'owner':
+            flash("You can't edit owner", category='danger')
+            return redirect('/admin/users')
 
         # Username, name, email
         username = request.form.get('username')
@@ -68,7 +78,7 @@ def admin_user_edit_page(id):
         if confirmation == 'True':
             if user.is_confirmed == False:
                 user.is_confirmed = True
-                user.confirmed_on = datetime.now() 
+                user.confirmed_on = str(datetime.now())[:19]
         if confirmation == 'False':
             if user.is_confirmed == True:
                 user.is_confirmed = False
@@ -76,6 +86,11 @@ def admin_user_edit_page(id):
 
         # Role
         role = request.form.get('role')
+        
+        if current_user.role != 'owner' and role == 'owner':
+            flash("You can't change owner role", category='danger')
+            return redirect(f'/admin/user/edit/{id}')
+
         if role != user.role:
             user.role = role
 
@@ -91,14 +106,18 @@ def admin_user_delete_page(id):
     form = SubmitForm()
     user = User.query.get(id)
     if form.validate_on_submit():
+        if user.role == 'owner' and current_user.role != 'owner':
+            flash("You can't delete owner", category='danger')
+            return redirect('/admin/users')
+        
         db.session.delete(user)
         db.session.commit()
         flash("User deleted successfully", category='success')
         return redirect('/admin/users')
     return render_template('Admin/delete-user.html', form=form, user=user)
 
+
 @app.route('/admin/products', methods=['GET','POST'])
 @admin_required
 def admin_products_page():
-    products = Product.query.all()
-    return render_template('Admin/products.html', products=products)
+    pass
